Load Regular Expressions From Settings File The regular expressions in the settings file are loaded into a list. There can be any number of these. They are then applied to each string to be renamed. Unfortunately, the order isn't guaranteed. This could probably be fixed by having 'override' be a list of items (dicts), instead of one giant dict. That might preserve the order.
Override is bad name. That should be changed.

Another unfortunate issue is the escaping that has to take place in settings
file. To replace a backslash, the string "\\\\" must be present. I think it
needs to be double escaped, though I don't quite understand why. The default
one, in the script, is just r'\\'.